+++ easier68k/assembler/assembler.py
@@ -11,13 +11,6 @@ from ..core.util.find_module import find_opcode_cls
 # noinspection PyUnresolvedReferences
 from ..core.opcodes import *
 
-valid_opcodes = [
-    'easier68k.core.opcodes.move',
-    'easier68k.core.opcodes.dc',
-    'easier68k.core.opcodes.lea',
-    'easier68k.core.opcodes.simhalt'
-]
-
 MAX_MEMORY_LOCATION = 16777216  # 2^24
